博客删除服务器逻辑完善

+++ BlogSystem/src/main/java/com/learn/controler/blog/DeleteBlogServlet.java
@@ -1,12 +1,15 @@
 package com.learn.controler.blog;
 
 import com.learn.dao.BlogDao;
+import com.learn.pojo.Blog;
+import com.learn.pojo.User;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
 @WebServlet("/deleteBlog")
@@ -15,6 +18,22 @@ public class DeleteBlogServlet extends HttpServlet {
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         //设置字符编码集
         req.setCharacterEncoding("utf8");
+
+        //服务器这里再次进行用户是否登录判断
+        HttpSession session = req.getSession(false);
+        if(session == null){
+            resp.setContentType("text/html; charset=utf8");
+            resp.getWriter().write("用户未登录!");
+            return;
+        }
+        //获取session中的user属性
+        User user = (User) session.getAttribute("user");
+        if(user == null){
+            resp.setContentType("text/html; charset=utf8");
+            resp.getWriter().write("用户未登录!");
+            return;
+        }
+
         //获取 博客id
         String blogId = req.getParameter("blogId");
         //先判断接收的参数是否合法
@@ -23,8 +42,17 @@ public class DeleteBlogServlet extends HttpServlet {
             resp.getWriter().write("blogId非法!");
             return;
         }
-        //通过博客Id来删除具体的博客
+
         BlogDao blogDao = new BlogDao();
+        //再次判断用户登录的 username 和 博客作者 是否一致
+        Blog blog = blogDao.selectBlogId(Integer.parseInt(blogId));
+        if(blog.getUserId() != user.getUserId()){
+            resp.setContentType("text/html; charset=utf8");
+            resp.getWriter().write("没有权限删除此博客!");
+            return;
+        }
+
+        //通过博客Id来删除具体的博客
         int id = Integer.parseInt(blogId);
         boolean result = blogDao.deleteBlogId(id);
         if(result){
